+++ source/extensions/omni.isaac.examples/omni/isaac/examples/franka_cortex/franka_cortex.py
@@ -12,11 +12,11 @@ import carb
 import numpy as np
 import omni
 from isaacsim.core.api.objects import DynamicCuboid, VisualCuboid
-from omni.isaac.cortex.cortex_utils import load_behavior_module
-from omni.isaac.cortex.cortex_world import Behavior, CortexWorld, LogicalStateMonitor
-from omni.isaac.cortex.dfb import DfDiagnosticsMonitor
-from omni.isaac.cortex.robot import CortexFranka, add_franka_to_stage
-from omni.isaac.cortex.tools import SteadyRate
+from isaacsim.cortex.framework.cortex_utils import load_behavior_module
+from isaacsim.cortex.framework.cortex_world import Behavior, CortexWorld, LogicalStateMonitor
+from isaacsim.cortex.framework.dfb import DfDiagnosticsMonitor
+from isaacsim.cortex.framework.robot import CortexFranka, add_franka_to_stage
+from isaacsim.cortex.framework.tools import SteadyRate
 from omni.isaac.examples.cortex.cortex_base import CortexBase