+++ source/extensions/omni.isaac.manipulators/python/grippers/surface_gripper.py
@@ -11,8 +11,8 @@ import numpy as np
 import omni.kit.app
 from isaacsim.core.api.articulations.articulation import Articulation
 from isaacsim.core.api.utils.types import ArticulationAction
+from isaacsim.robot.surface_gripper._surface_gripper import Surface_Gripper, Surface_Gripper_Properties
 from omni.isaac.manipulators.grippers.gripper import Gripper
-from omni.isaac.surface_gripper._surface_gripper import Surface_Gripper, Surface_Gripper_Properties
 
 
 class SurfaceGripper(Gripper):