+++ source/standalone_examples/api/omni.isaac.cortex/example_cortex_sync_belief_main.py
@@ -29,10 +29,10 @@ import numpy as np
 from behaviors.franka.franka_behaviors import ContextStateMonitor, behaviors
 from isaacsim.core.api.objects import DynamicCuboid
 from isaacsim.core.api.utils.extensions import enable_extension
-from omni.isaac.cortex.cortex_object import CortexObject
-from omni.isaac.cortex.cortex_utils import load_behavior_module
-from omni.isaac.cortex.cortex_world import CortexWorld
-from omni.isaac.cortex.robot import add_franka_to_stage
+from isaacsim.cortex.framework.cortex_object import CortexObject
+from isaacsim.cortex.framework.cortex_utils import load_behavior_module
+from isaacsim.cortex.framework.cortex_world import CortexWorld
+from isaacsim.cortex.framework.robot import add_franka_to_stage
 
 enable_extension("omni.isaac.cortex_sync")
 from omni.isaac.cortex_sync.cortex_ros import CortexControlRos, CortexObjectsRos, cortex_init_ros_node